// src/position/position.enum.ts
export enum Position {
  GetPosition = 'GET_POSITION',
  CreateAndModifyPosition = 'CREATE_AND_MODIFY_POSITION',
  DeletePosition = 'DELETE_POSITION',

  GetEmployee = 'GET_EMPLOYEE',
  AddEmployee = 'ADD_EMPLOYEE',
  FireEmployee = 'FIRE_EMPLOYEE',
  GrantEmployeePosition = 'GRANT_EMPLOYEE_POSITION',

  GetBranch = 'GET_BRANCH',
  CreateAndModifyBranch = 'ADD_AND_MODIFY_BRANCH',
  DeleteBranch = 'DELETE_BRANCH',
  AddEmployeeToBranch = 'ADD_EMPLOYEE_TO_BRANCH',
  RemoveEmployeeFromBranch = 'REMOVE_EMPLOYEE_FROM_BRANCH',
}